+++ Source/Examples/ExampleLibrary/Series/HistogramSeriesExamples.cs
@@ -29,6 +29,13 @@ namespace ExampleLibrary
             return CreateExponentialDistribution();
         }
 
+        [Example("Exponential Distribution (logarithmic)")]
+        [DocumentationExample("Series/HistogramSeries")]
+        public static PlotModel ExponentialDistributionLogarithmicAxis()
+        {
+            return CreateExponentialDistribution(true);
+        }
+
         [Example("Label Placement")]
         public static PlotModel HistogramLabelPlacement()
         {
@@ -122,10 +129,13 @@ namespace ExampleLibrary
             public string Description { get; }
         }
 
-        public static PlotModel CreateExponentialDistribution(double mean = 1, int n = 10000)
+        public static PlotModel CreateExponentialDistribution(bool logarithmicYAxis = false, double mean = 1, int n = 10000)
         {
-            var model = new PlotModel { Title = "Exponential Distribution", Subtitle = "Uniformly distributed bins (" + n + " samples)" };
-            model.Axes.Add(new LinearAxis { Position = AxisPosition.Left, Title = "Frequency" });
+            var model = new PlotModel { Title = logarithmicYAxis ? "Exponential Distribution (logarithmic)" : "Exponential Distribution", Subtitle = "Uniformly distributed bins (" + n + " samples)" };
+            model.Axes.Add(
+                logarithmicYAxis ? 
+                    (Axis)new LogarithmicAxis { Position = AxisPosition.Left, Title = "Frequency"} : 
+                    new LinearAxis { Position = AxisPosition.Left, Title = "Frequency" });
             model.Axes.Add(new LinearAxis { Position = AxisPosition.Bottom, Title = "x" });
 
             Random rnd = new Random(1);